+++ src/main/java/de/markusbordihn/lobby/commands/GamingCommand.java
@@ -40,9 +40,6 @@ import de.markusbordihn.lobby.teleporter.PlayerTeleportManager;
 
 public class GamingCommand extends CustomCommand {
 
-  public static final String DIMENSION_NAME = "Gaming";
-  public static final int PERMISSION_LEVEL = 0;
-
   private static final CommonConfig.Config COMMON = CommonConfig.COMMON;
 
   private static Map<Player, Long> coolDownPlayerMap = new ConcurrentHashMap<>();
@@ -53,9 +50,11 @@ public class GamingCommand extends CustomCommand {
     if (Boolean.FALSE.equals(COMMON.gamingEnabled.get())) {
       return;
     }
-    registerCommand(COMMON.gamingCommandName.get(), DIMENSION_NAME, PERMISSION_LEVEL);
+    registerCommand(COMMON.gamingCommandName.get(), COMMON.gamingDimensionName.get(),
+        COMMON.gamingCommandPermissionLevel.get());
     dispatcher.register(Commands.literal(COMMON.gamingCommandName.get())
-        .requires(cs -> cs.hasPermission(PERMISSION_LEVEL)).executes(command));
+        .requires(cs -> cs.hasPermission(COMMON.gamingCommandPermissionLevel.get()))
+        .executes(command));
   }
 
   @Override
@@ -66,8 +65,10 @@ public class GamingCommand extends CustomCommand {
     Long coolDownTimer = coolDownPlayerMap.getOrDefault(player, null);
     Long currentTimer = java.time.Instant.now().getEpochSecond();
     if (coolDownTimer != null && coolDownTimer > currentTimer) {
-      sendFeedback(context, Component.translatable(Constants.TELEPORT_FAILED_COOLDOWN,
-          DIMENSION_NAME, coolDownTimer - currentTimer).withStyle(ChatFormatting.RED));
+      sendFeedback(context,
+          Component.translatable(Constants.TELEPORT_FAILED_COOLDOWN,
+              COMMON.gamingDimensionName.get(), coolDownTimer - currentTimer)
+              .withStyle(ChatFormatting.RED));
       return 0;
     } else {
       coolDownPlayerMap.put(player, currentTimer + COMMON.generalCommandCoolDown.get());
@@ -76,24 +77,26 @@ public class GamingCommand extends CustomCommand {
     // Provide feedback to the player for their teleporter request.
     if (DimensionManager.getGamingDimension() == null) {
       sendFeedback(context, Component.translatable(Constants.UNABLE_TO_TELEPORT_MESSAGE,
-          DIMENSION_NAME, DimensionManager.getGamingDimensionName()));
+          COMMON.gamingDimensionName.get(), DimensionManager.getGamingDimensionName()));
     } else if (Boolean.TRUE.equals(!COMMON.gamingRestrictCommand.get())
         || player.getLevel() != DimensionManager.getGamingDimension()) {
       if (Boolean.TRUE.equals(COMMON.teleportDelayEnabled.get())
           && COMMON.teleportDelayCounter.get() > 0) {
-        sendFeedback(context, Component.translatable(Constants.TELEPORT_TO_IN_MESSAGE,
-            DIMENSION_NAME, COMMON.teleportDelayCounter.get()).withStyle(ChatFormatting.GREEN));
+        sendFeedback(context,
+            Component.translatable(Constants.TELEPORT_TO_IN_MESSAGE,
+                COMMON.gamingDimensionName.get(), COMMON.teleportDelayCounter.get())
+                .withStyle(ChatFormatting.GREEN));
         PlayerTeleportManager.teleportPlayerToGaming(player);
       } else {
-        sendFeedback(context, Component.translatable(Constants.TELEPORT_TO_MESSAGE, DIMENSION_NAME)
-            .withStyle(ChatFormatting.GREEN));
+        sendFeedback(context,
+            Component.translatable(Constants.TELEPORT_TO_MESSAGE, COMMON.gamingDimensionName.get())
+                .withStyle(ChatFormatting.GREEN));
         DimensionManager.teleportToGaming(player);
       }
     } else {
       sendFeedback(context,
-          Component
-              .translatable(Constants.TELEPORT_FAILED_ALREADY_IN_DIMENSION_MESSAGE, DIMENSION_NAME)
-              .withStyle(ChatFormatting.YELLOW));
+          Component.translatable(Constants.TELEPORT_FAILED_ALREADY_IN_DIMENSION_MESSAGE,
+              COMMON.gamingDimensionName.get()).withStyle(ChatFormatting.YELLOW));
     }
     return 0;
   }